Time: 208 ms (30.35%) | Memory: 0B (100.00%) - LeetSync

+++ 1452-restaurant-growth/restaurant-growth.sql
@@ -7,9 +7,14 @@ WITH cte AS (
     GROUP BY visited_on
     ORDER BY visited_on
 )
-SELECT 
+
+SELECT
     visited_on,
-    SUM(amount) OVER(ROWS BETWEEN 6 PRECEDING AND CURRENT ROW) AS amount,
-    ROUND(AVG(amount::DECIMAL) OVER(ROWS BETWEEN 6 PRECEDING AND CURRENT ROW), 2) AS average_amount 
+    SUM(amount) OVER(
+        ROWS BETWEEN 6 PRECEDING AND CURRENT ROW
+    ) AS amount,
+    ROUND(AVG(amount) OVER(
+        ROWS BETWEEN 6 PRECEDING AND CURRENT ROW
+    ), 2) AS average_amount
 FROM cte
 OFFSET 6
